feat: move dateTimeToShift utility function to a separate file
